Fixed file bind-mounting issues in Docker. Added Travis CI.
