+++ .gitignore
@@ -1,3 +1,4 @@
+.ash_history
 .config
 .env
 .npm
