+++ docker/Dockerfile-app.dev
@@ -1,7 +1,5 @@
 FROM node:10-alpine
 
-EXPOSE 3000
-
 RUN mkdir -p /home/app/app
 ENV HOME=/home/app/app
 WORKDIR ${HOME}
